proposal_finder.go proposal_searcher.go proposal_updater.go json_output.go proposal_finder.go proposal_finder_test.go proposal_merger.go proposal_searcher.go proposal_updater.go
