+++ internal/forge/gh/proposal_searcher.go
@@ -6,6 +6,8 @@ import (
 	. "github.com/git-town/git-town/v21/pkg/prelude"
 )
 
+var _ forgedomain.ProposalSearcher = ghConnector
+
 func (self Connector) SearchProposal(branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
 	out, err := self.Backend.Query("gh", "--head="+branch.String(), "--json=number,title,body,mergeable,headRefName,baseRefName,url")
 	if err != nil {
