+++ internal/forge/glab/proposal_merger.go
@@ -3,8 +3,11 @@ package glab
 import (
 	"strconv"
 
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
 	"github.com/git-town/git-town/v21/internal/git/gitdomain"
 )
+
+var _ forgedomain.ProposalMerger = glabConnector
 
 func (self Connector) SquashMergeProposal(number int, message gitdomain.CommitMessage) error {
 	return self.Frontend.Run("glab", "mr", "merge", "--squash", "--body="+message.String(), strconv.Itoa(number))
