+++ internal/forge/glab/proposal_searcher.go
@@ -6,6 +6,8 @@ import (
 	. "github.com/git-town/git-town/v21/pkg/prelude"
 )
 
+var _ forgedomain.ProposalSearcher = glabConnector
+
 func (self Connector) SearchProposal(branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
 	out, err := self.Backend.Query("glab", "--source-branch="+branch.String(), "--output=json")
 	if err != nil {
